Move user-facing messages out of Uploader
The Uploader class contains the logic for talking to the API and sending
files to Permanent, and for sending status events to listeners for the
sake of observability. It should not have any kind of user-facing code
or strings; that responsibility lies upstream in the call stack.

Make the UploadService class, which already contains some calls to the
MessageService to flash UI messages, watch for the relevant Uploader
events and show the UI messages that previously Uploader was showing.

(I could certainly be persuaded that UploadService is also not the right
place for this, but this is an incremental improvement and unblocks
further changes to the Uploader.)

PER-8029 Mdot changes to call the new API end points