Refactor to make Uploader injectable
Previously, the Uploader class was directly constructed by its sole
user, UploadService. That works well enough for the simple use case
where Uploader has a subset of UploadService's dependencies. However, in
order for Uploader to implement the new direct-to-S3 upload workflow, it
will need to be able to make non-Permanent-API HTTP requests, which in
turn will require a new dependency.

Wire up the Uploader class to be injectable via Angular's dependency
injection[1].

As this is purely a refactor, there should be no change in behavior; it
just sets up future changes to be made more easily.

[1] https://angular.io/guide/dependency-injection