Use enum names instead of indices
The upload progress component examines the session status to determine
what content to show. Angular limits the data visible to an HTML view to
its corresponding class; anything outside of that class is not
accessible in the view.

While that limitation makes sense in general, in this specific case, it
led to using the numerical indices of the `UploadSessionStatus` enum
instead of its names. That makes the view code harder to read, and it
makes changing the number or order of the values in the enum dangerous
in a way that is hard to discover.

The workaround is to assign the enum to a member variable of the
component class. This makes the enum names available to the HTML view.

See also: https://stackoverflow.com/a/42464835