Refactor fileUploadComplete into unified progress
Start the migration from a wide (and, for me, hard to think about)
variety of ways to get insight into the status of uploading a batch of
files into a single, unified stream of events.

Replace the `fileUploadComplete` event emitter with a new emitter and a
new type. Update the sole subscriber to that old emitter to instead
subscribe to the new one.