Refactor uploadItem into unified progress
Continue the migration from a wide (and, for me, hard to think about)
variety of ways to get insight into the status of uploading a batch of
files into a single, unified stream of events.

Replace the `uploadItem` event emitter with the unified progress
emitter. Update the sole subscriber to the old emitter to instead
subscribe to the new one.

Since we're breaking the direct link between the current upload item and
the progress meter, we need to emit events more frequently to make up
for it: each chunk callback needs to emit its progress. Relatedly, the
event emitted on completing a file upload should happen before our
recursive iteration; I think what was happening was that a slew of "file
done!" events were emitted at the very end, in reverse order. Send out
progress events more intentionally throughout the process of uploading
an individual file, and send the completed file event before moving on
to the next file.