Refactor uploadSessionStatus into unified progress
Continue the migration from a wide (and, for me, hard to think about)
variety of ways to get insight into the status of uploading a batch of
files into a single, unified stream of events.

Replace the `uploadSessionStatus` event emitter with the unified
progress emitter. Update the subscribers to the old emitter to
instead subscribe to the new one.

Note that one event was deleted entirely: the one that emits just "in
progress". Instead, the other progress events that are emitted have
their session status set to "in progress".