Merge subscription into constructor
With the previously-separate uploadSessionStatus and uploadItem event
streams merged into a single progress stream, it does not make sense to
have multiple subscriptions.

Additionally, all of the work being done in ngOnInit can be done equally
well in the constructor; none of it requires the rest of the component
to be initalized, and in particular nothing in the DOM.

Merge the separate subscription that was being done in ngOnInit into the
subscription set up in the constructor, as well as the reference to the
uploader's file count.