Refactor file count into upload progress event
Continue the migration from a wide (and, for me, hard to think about)
variety of ways to get insight into the status of uploading a batch of
files into a single, unified stream of events.

The last bit of world-visible state in the uploader was the count of
files during this upload session. The progress component displays this
information ("2 of 4", or "2 uploaded, 2 skipped" in an error), and was
previously reaching into the class to get it.

Refactor how this information is shared by making the member variable
private and including a statistics section in the upload progress event.

This concludes the upload progress visibility refactoring, and as a
whole, will allow us to radically restructure the uploader without
making any further changes to unrelated components: they all use this
single, unified API.