When `dateOnly=true` use plain YYYY-MM-DD values
When time is hidden in all current uses, the value chosen is meant to be used in a SQL `date` type column, which doesn't allow for any time informaton of any kind. So, if we only output the `YYYY-MM-DD` format anytime this flag is true, this output can safely be inserted into a `date` column.