Refresh the file list after each successful upload
Before this change, we would only refresh the file list if *all* the
files previously queued for upload into this folder were uploaded
successfully. This made sense for the old upload workflow, where we
upload files as a batch, and create placeholders for the to-be-uploaded
files.

In the new upload flow, we upload files one-at-a-time. Uploads can fail
partway through a batch, and we do still want to see the successfully
uploaded files in the file list. Remove the complicated calculations to
see if all the queued files in this folder were uploaded, and simply
refresh the file list when a file was successfully uploaded to this
folder.