Remove unused helper method
It seems that we haven't used the `debouncedRefresh` method since June.
Delete it.

In the somewhat unlikely event that we start having performance problems
due to too many refreshes during upload, to the tune of more than 1
upload every .75 seconds, we can revisit this and figure out if we need
debounce or throttle.