Upload directly to S3
Implement the new upload flow in the front-end.

The old upload flow was approximately:
- call postMetaBatch to pre-create the database entries for the
  metadata of the files to be uploaded
- build a websocket to the old uploader
- stream files over the websocket
- tear down the websocket, both on success and failure

The new upload flow is approximately:
- call new getPresignedUrl API endpoint
- upload directly to S3
- call registerRecord with file metadata

Progress Notifications
----------------------

In both flows, we need to send out progress notifications to other
components that may be listening. Previous work has structured the
notifications in the old flow to be easy to convert to the new flow.

The Angular HttpClient interface for POST requests[1] allows us to
request periodic callback with the progress of the request; this is
currently listed in the documentation as "Overload #7":

    Constructs a POST request that interprets the body as a JSON object
    and returns the full event stream.

Each event in the stream is of type HttpEvent[2], which has a key `type`
whose value is of type HttpEventType[3]. The file is finished uploading
when we get a `Response` from S3.

One nuance here is that the "upload done" message must be sent after the
call to registerRecord, and not merely after the file is successfully
uploaded to S3, because that message triggers the client to refresh the
directory. If the refresh happens before the file is registered with the
API, then the refresh doesn't find anything new, and it looks like the
file didn't upload; the user would then need to manually refresh the
page to pick up the newly uploaded file.

File Types
----------

The new upload flow requires a file type to generate the pre-signed
request to S3, and in uploading the file to S3. Unfortunately, sometimes
the browser doesn't provide a file type. That seems to be the case when
dragging-and-dropping a folder of files in Firefox, as well as for some
file types (such as .log) in Chrome on macOS.

RFC 2046[4] defines a MIME type (now media type) for such an unknown
case:
        "application/octet-stream" is to be used in the
        case of uninterpreted binary data

Use this as a default value for when the file type is either empty or
missing.

This is primarily to make S3 happy: S3 requires the content type to be
specified in the request, and so we need to include it while generating
the presigned request. Later, during post-upload processing, the backend
does additional file type detection, and eventually sets the file's
content-type with what the server detects.

BinaryJS Dependency
-------------------

The old upload flow required connecting via a websocket to stream files
to the old uploader. This was handled on the client side via the library
BinaryJS[5].

This dependency was one of the motivating factors for changing how we
did uploads; a small one, to be sure, but a part of the decision. It
hasn't been updated in more than half a decade, it doesn't have any
upstream type definitions, and we had vendored it in instead of using
npm.

With the conversion to the new upload flow, we no longer need this
dependency, so delete it and our in-house type definitions for it.

[1] https://angular.io/api/common/http/HttpClient#post
[2] https://angular.io/api/common/http/HttpEvent
[3] https://angular.io/api/common/http/HttpEventType
[4] https://tools.ietf.org/html/rfc2046
[5] https://github.com/binaryjs/binaryjs

PER-8029 Mdot changes to call the new API end points