Add upload session management
Part of the complexity of the old uploader was in managing the queue.
This was not (only?) required by the binary websocket library; it turns
out to be an essential part of the application.

There are two triggering use cases: when a folder full of folders full
of files is drag-and-dropped onto the window, which gets mapped to
multiple (folder, files[]); and when a file or folder is uploaded before
the previous upload session is complete.

The failure mode to not managing the queue was some UI weirdness:
multiple queues would be uploaded in parallel, and the progress bar
would flicker between them as updates from each came in. The upload
functionality worked correctly, but the UI was confusing at best.

Make upload sessions a first class concept, and make it explicit that we
are queuing files to be uploaded. Move the queue management out of
uploader and into a new upload session class, which allows each to be
more focused.

We looked at using RxJS Observables throughout, instead of passing
callbacks, but the code we prototyped turned out to be more complex and
harder to understand. Async/await felt like the better fit for this
problem.

PER-8029 Mdot changes to call the new API end points

Co-authored-by: Andrew Atwood <aatwood@permanent.org>