WIP: Handle verification clicks
When the user clicks the verification link in their email, take them
to the verification page with the code prefilled (instead of
erroneously redirecting to the login page).  This also requires a
change in the `api` repo.

Currently this does not work if the user is not logged into Permanent
(with some account).  Is that what we want?  These edits don't change
that.

The still-failing part is this case:
- My account cecilia+testverify2 is unverified
- I'm logged in as cecilia+local
- I click the verification link in email for cecilia+testverify2
- web-app should log me out of cecilia+local and fill out the form for
  cecilia+testverify
- Currently it fails to create an account var for cecilia+testverify
  and I am going to take a break now instead of looking into why.