Refresh file list on upload completion
When addressing the multiple-uploads issue by adding upload sessions in
commit e160aa0b1ca0a768d8d1844e7f9c4ff0e046e49f, we missed a bit of
accounting, and stopped sending upload complete messages for individual
items in the queue. The upload progress UI worked fine, because it is
not watching for those messages, but the upload services triggers a
refresh of the file list on that message.

Emit a progress event for completing an upload, so that the file list
refresh is triggered correctly.

PER-8029 Mdot changes to call the new API end points