Minimal change to make verification links work
Along with the paired change to `api`, this is the minimal change to
fix PER-7970.  Possibly we should just stop here.

The next step would be to fix the commented-out code managing the case
where we have a session for a *different* account from the one with an
email being verified.