Upload next in queue regardless of success
This fixes one oversight that's to blame for disappearing upload sessions. As written, when an item successfully uploads, the queue is triggered to continue. If it fails, however, nothing happens, and the session is left in an `inProgress` state with the queue abandoned. Future attempted uploads fail even though they are added to the queue because nothing is triggered the next upload from the queue. So uploads remain broken until refresh. Not great!

This change makes sure that the next upload in the queue is always attempted, regardless of success or failure, which will continue until there are no more items in the queue, at which point the session resets, and any future uploads are free to start a new session.