Show upload progress bar during folder creation
When dragging and dropping, the first UploadSessionStatus event wasn't
being emitted until the very first file actually begins uploading. For a
flat list of files, this doesn't cause any issues, but when dragging any
number of folders or subfolders, there would be a period of zero UI
feedback while the API calls to create the needed folders was executing
before there was any indication that the upload was about to begin.

Send the `Start` progress event earlier in the process, so that the
"Connecting..." message is shown while folders are created. The calling
UploadService needs to call this new method sooner, which makes the API
slightly more complex with some temporal coupling, but we fall back to
sending the `Start` event when files are first queued for upload if
`startSession` was not previously called.

Now that we have another bit of state to track (whether or not we've
sent the `Start` event), and have a `startSession` method, it makes
sense to move the end-of-session accounting out into its own,
complementarily-named helper method.

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>