PER-8261: Make tags case sensitive
Tags need to be case sensitive to match Library of Congress standards,
and specifically for a partner of ours.  The tag search by default was
case insensitive, but "Fuse" has an option for case sensitivity, so
this turns that on.  See https://fusejs.io/api/options.html if you're
interested.