Create `load-secrets` script
At app runtime, Angular is run exclusively client-side and has no access to `process.env`, so any secrets must be accessible in a static file available at build time.

To facilitate this, `load-secrets.js` uses `dotenv` to load environment variables from the `.env` file, and in turn writes them to the file `src/secrets.ts`. This file is added to `.gitignore` to prevent any leakage via source control, but is available to the TypeScript compiler during build as a static file containing the secrets from .env in plain Javascript object format.

`load-secrets` must be run prior to any dev server or build to ensure that the `secrets.ts` file is up to date. When running, it checks against `required-secrets` to see if any variables listed as required are missing, and terminates the process if so, which terminates the build step.

`SecretsService` reads from `secrets.ts` and re-validates the available via the `required-secrets` list. Secrets are accessed via the `get` method, which throws an exception if the secret is undefined to minimize downstream errors due to missing secrets.