Move all menu contents into scroll element
Previously, on desktop, everything above My Files did not scroll on smaller screens, only the navigable sections of the App (My Files, Shares, etc.). On smaller screens and certain zoom levels, especially when the "marquee" of archive options was expanded, this would lead to a tiny scrolling box with only one section visible at a time, which was confusing and difficult to navigate.

To simplify the DOM structure of the left-hand menu, now the entire menu, below the Permanent.org logo, will scroll as needed on smaller screens and on mobile.