Simplify resolve for Apps root folder
Rather than check for ChildItemVOs on the folder for any given integration to determine if it it's been run before, we'll assume that if the folder exists at all that the integration has been run. Since we don't care what's in the folder, we can greatly simplify the API calls needed to load the Apps root screen down to just one, and also simplify the ConnectorComponent to remove the `hasFiles` boolean.