Style shepherd.js
This removes the compiled CSS from the angular.json `styles` array and copies the base Shepherd styles into our scss, as they make it a little difficult to override. Because of this, I manually removed the included button styling in favor of the button styles already set, and override a few other properties.