PER-7945: Show donation form in storage modal
This is a very intermediate step toward making donations work for a
logged-in user without leaving the app.  At this point the donation
form appears but I would not expect it to work.  These changes also
would also probably break the regular donation form, so should not
live as is!  I just wanted to record this point.