Use get instead of getByArchiveNbr
I removed getByArchiveNbr as part of the Etherpad work, thinking that
it was never used and was a duplicate of archive/get.  It was used,
though, by web-app, to resolve public archives.  After pushing out the
back-end change, public archives broke.  This fixes them by changing
the call to use archive/get instead.  There will be some followup
cleanup.