PER-8363: Onboarding Front End
This PR adds the OnboardingModule and the basic UI components associated with it.
While it isn't fully wired up to the process yet, the WIP onboarding page can
still be accessed by going to /app/onboarding in the web-app.

A debugger component is also included in this version of the front end.