PER-8331: Keep file hidden after move is queued
This PR keeps a file hidden after a move operation is queued.
Because the move endpoint only queues up a file to be moved,
and there are no other endpoints or methods to check this,
we cannot confirm when the file has actually been moved. With this PR,
the Data Service continues to hide the file through the initial refresh
triggered by the UI so that the moved file does not re-appear
immediately after it was moved.

This does create other edge-case-y bugs since this is a workaround.
If a move operation is successfully queued but then rolled back after
the fact, the UI not only communicates that the move was successful,
but doubles down on the lie by pretending like it has been moved.
Not ideal, but it is a bug with a much smaller scope than the original
bug which happens on literally any move operation.