Remove unused fields from AccountVO
As part of some other work[1] on the account table, I'm removing several
columns that were never used. Also remove them from the types and the
saved responses used for testing, so that web-app stays in sync with the
back-end.

(Note that while the back-end PR[1] also adds a column, it does not yet
add fields to AccountVO, so the new column should not show up in
responses - and might never do so.)

[1] https://bitbucket.org/permanent-org/back-end/pull-requests/107

PER-8522 Transfer user credentials to auth vendor
PER-8523 Validate auth tokens on API requests