PER-8248: Show navigation style sooner in process
This PR fixes a bug where the UI sometimes does not immediately indicate 
its navigating to another folder when clicking into one. This bug is 
caused by the front end needing to fetch more information on that folder 
before being able to navigate to it, and until that information is 
actually fully fetched the UI does not actually indicate is loading or 
navigating. The code additions here trigger the UI to enter the greyed 
out "loading" state right when it fetches that additional information.