Fix dependency injection issues on public galleries.
There was initially an error with the full screen file viewer on public galleries as a result of adding the metadata editing functionality, the EditService was suddenly a needed dependency and since no other components in the Public side of the app provided it, there was a dependency injection error. This fix makes sure that the EditService is properly provided in the public side of the app (even if it's never actually used).