Fix AccountService null issues
I decided to remove the archives.length check on one call because upon
further inspection, it doesn't matter if an empty array is passed in.
This check was mainly an artifact of me originally thinking the
AccountService needed a special flag to know when it was in the
onboarding state.

As for the other check, I've added an optional chain operator so as to
avoid any null access errors.