Only allow new onboarding flow from main signup
There are half a dozen different ways to sign up for an account. For the
sake of simplicity and speed, for the moment we'll only allow opting in
to the new onboarding workflow when signing up from the main page
(/app/auth/signup).

The account service and API adapter should not need to know the rules
for when we do and do not want to automatically create a default
archive, so instead update them to accept a parameter. Set the secondary
signup forms (from the newsletter, from the claim storage page, and from the
share preview) to always create a default archive. The primary signup
page will continue to examine the query parameters, with `createArchive`
and `noArchive` continuing to override the default behavior.

We now always send a `createArchive` argument to the API, rather than
relying on its default behavior when the SimpleVO is absent.

PER-8725 Start A/B testing signups