Create default archive in old share signups
We want to opt out of the new onboarding workflow when the signup is in
response to a file/folder share, I think because such shares need a
target archive. We are mainly doing that already, by always passing
`true` in SharePreviewComponent.

There is also an old style of signup link, which includes `shid` and
`tp` query parameters. I believe these have not been used since the old
web-client, but the functionality still exists, so let's be sure to
cover it. Always create a default archive for such old share links.

PER-8725 Start A/B testing signups