Randomly (never) opt into the new onboarding flow
Create a constant for the chance for a given signup to use the new
onboarding flow instead of the old onboarding flow, and randomly choose
whether or not to create a default archive based on that chance.

For now, set the chance to 0, so that the old onboarding flow is always
chosen in the absence of a query parameter. That allows us to deploy
this all the way out to prod, and to easily tune this number later.

PER-8725 Start A/B testing signups