Remove deprecated ng serve options
In Angular 11, baseHref and deployUrl are not needed in the ng serve
configuration. deployUrl is used in the browser builder configuration
and we don't need it in serve as well. baseHref is only needed in the
local proxy browser builder configuration so move it there from the
serve configuration.

PER-8819: Upgrade Angular

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>