Disable legacy relativeLinkResolution
Angular 11 changed the default value of relativeLinkResolution from
'legacy' to 'corrected'. This changes the evaluation of links that begin
with './'. Our application doesn't have any such links, so we should be
able to use the new 'corrected' behavior. The behavior of '../' links
appears to be unchanged, so no changes need to be made as a result.
Delete the override introduced by the migration to Angular 11.

See also: https://v11.angular.io/api/router/ExtraOptions#relativeLinkResolution

PER-8819: Upgrade Angular

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>