Remove ngbTooltip errors from tests
Something about the upgrade to Angular 11 caused these errors to appear
in tests. The cause of these errors is that the proper ng-bootstrap
module isn't being imported properly in the testing environment. (This
is definitely one of those "how did this ever work?" kind of issues).
This commit includes the proper module in the TestBed config of the
tests with these errors.

PER-8819: Upgrade Angular