Create Announcement Module and Component
This is the output from running the Angular CLI to generate the module
and component boilerplate for the new component. This component will be
getting its own module because it needs to be supported in many
different places: the Auth pages, the Core web app, and the Public
galleries. Putting the AnnouncementComponent in the SharedModule or
CoreModule (where we tend to put most commonly used components) would
also mean importing all the other components in the module. Since
AnnouncementModule isn't going to depend on anything itself, there's no
reason to pack it in the giant modules we have here. Let's put it in its
own module to make it easier to isolate and import elsewhere without
additional baggage.

PER-8875: Show warning in app about scheduled downtime