Basic functionality for AnnouncementComponent
This commit contains the basic functionality and tests for the
AnnouncementComponent. The component is built to take an Array of
`AnnouncementEvent`s and decide to show the first active announcement it
can find. It can be dismissed and it will remember this via
localStorage. This memory only applies to the currently active
announcement, so if in the future we need to show another announcement,
the new announcement will still be dismissed even if the localStorage
flag is set.

PER-8875: Show warning in app about scheduled downtime