Style and include AnnouncementComponent
This commit styles the AnnouncementComponent and also includes it in
various parts of the web-app. Due to the way that the web-app layout is
put together, simply including the extra height of the announcement
banner ended up breaking both the main app as well as the public gallery
page. As a result, this component does some things that may or may not
be seen as hacky: it manually searches for HTML elements I've marked as
"needing adjustment for the AnnouncementComponent" and adjusts their
spacing so that the layout isn't broken. It re-adjusts these styles back
to their default once the announcement is dismissed. I thought this
would be a better solution than those specific components having to know
and detect the AnnouncementComponent's presence.

PER-8875: Show warning in app about scheduled downtime