Remember to keep submenus open on left menu
This commit stores the state of the different left menu submenus in
SessionStorage so that they persist throughout page reloads and
refreshes. Especially with the redesign, it feels odd that they close
when moving between pages.

This also includes a quick change to not show the dropdown toggle if
there are no Apps subfolders.

PER-8667: Improvements to left side menu