Install Prettier
So far, all of the code in this repository has been manually formatted.
Instead of wasting time doing that, let's install Prettier to format our
code for us. This just installs Prettier as a dependency for our project
and does not include it in CI, or run it against the whole project.

web-app#91: Need Linting / Formatting Tool