Add support for optional secrets
Since we currently do not have any way of storing Featured Archive data
in the back-end, we need a way to store it somewhere else for the time
being. We've decided to store it in the `.env` file for now, but adding
this as a brand new secret might break things and we'd have to make
changes to everyone's local environments as well as the CI and
deployhment processes.

Instead let's plan to store the production data in the source code (for
now), and optionally require the FEATURED_ARCHIVES secret in `.env`,
defaulting to an empty array. This commit includes support for new
"optional" secrets with defined default values if they are not included
in the `.env` file.

PER-8683: Public gallery landing page (static)