Update Public Search Box Component for Gallery
The Public Gallery landing page features a search box to allow users to
search public archives. This component already exists in the public
archive pages, but needs to be customized to fit within the Public
Gallery page's design. This commit adds a new input to the Search Box
component that modifies its style if it's included on the Public Gallery
page.

This also deletes the test suite for this component which was previously
fully commented out, as well as exports the component from the
PublicModule so that the GalleryModule (or any other modules) can import
it.

PER-8683: Public gallery landing page (static)