Fix styling for nonprofit type archives
Although the `type.archive.nonprofit` archives are not really used
(there are probably some in production), they were originally given
conflicting CSS classes in the gallery. In case we make that an actual
archive type, or are just testing with weird data, we should make it
consistent and lump them in with Organizational archives.