Upgrade to Angular 12
This updates the core Angular libraries to Angular version 12. There are
also upgrades to various peer dependencies which need to be upgraded to
versions compatible with Angular 12, such as ng-bootstrap and
shallow-render.

This upgrade was performed with the following (long!) command:

```
npx ng update @angular-eslint/builder@12 \
@angular-eslint/eslint-plugin@12 \
@angular-eslint/eslint-plugin-template@12 \
@angular-eslint/schematics@12 \
@angular-eslint/template-parser@12 \
@angular/core@12 @angular/cli@12 @angular/cdk@12 \
@angular/google-maps@12 @angular/compiler-cli@12 \
@angular/language-service@12 @angular/localize@12 \
@angular/animations@12 @angular/common@12 \
@angular/forms@12 @angular/platform-browser@12 \
@angular/router@12 @ng-bootstrap/ng-bootstrap@10 \
shallow-render@12 @angular-devkit/build-angular@12
```

PER-8853: Upgrade to Angular 12