Fix invalid Promise typing
With Angular and its dependencies upgraded, the Promise on line 97 of
`publish.component.ts` creates a compilation error. This can be fixed by
specifying the type of the Promise with `Promise<void>` instead of
`Promise`.

Because of autoformatting, the other changes in this file are just
changes from Prettier.

PER-8853: Upgrade to Angular 12