Use currently maintained fork of vis-timeline
Previously we published our fork of vis-timeline as
`@permanent.org/vis-timeline`, but at a certain point we changed it to
`@permanentorg/vis-timeline`. However, we never changed which package
the web-app was using, so it has still been using the old package this
entire time. Let's use the proper package so when we merge dependabot
PRs in the vis-timeline fork, they actually appear in web-app.