Crop Public Gallery Landing Page Image
There is a lot of whitespace on this image that ends up disrupting the
layout. Let's get rid of it so we can show more of the image and things
line up with the perceived borders of the image instead of what looks
like empty white space.

PER-8683: Public Galler Landing Page (static)