Delete unused ArchiveSelectorComponent
Not to be confused with the ArchivePickerComponent, the
ArchiveSelectorComponent was used in earlier versions of the left menu.
In the current left menu design, an Archive Selector does not exist, so
this component is unused. Let's delete the component, and also delete
the now unused `showArchiveSelector` property on the left menu.

PER-8972: Delete unreachable components