Fix Public Gallery environment detection issue
A tricky thing we didn't catch in the original PR was that the
`archives` property on the GalleryComponent class was being intialized
before the `environment` property. As a result, when loading in Featured
Archives, the environment property was still undefined and could not
detect when the application was in production mode, making it unable to
load the production Featured Archives data. This fixes this issue by
explicitly defining the `environment` property in `getFeaturedArchives`
so this initialization issue doesn't happen.

I also fixed some CSS that doesn't cause any issues with archives but
does cause issues in the rare case there are no featured archives (when
the fixed bug is active or on the staging environment for example).