Add AndroidAppNotifyComponent
This component responds to the `beforeinstallprompt` event that is sent
out by Chrome on Android to notify users of the Android App. It is part
of the AnnouncementModule since it shares some code with the
AnnouncementComponent and also just acts similarly and should be placed
next to the pr-announcement component.

PER-8850: Notify members that they can download app (Android)