Upgrade to Angular 13
These are the npm dependency updates for Angular version 13. This
includes not only upgrades to the core Angular libraries to v13, but
also updates to various peer dependencies which needed to be upgraded as
well. (For example, ts-node was upgraded because Angular 13 also bumps
up the Typescript version).

This upgrade was in part performed with the following command:

```
npx ng update @angular-eslint/builder@13 \
@angular-eslint/eslint-plugin@13 \
@angular-eslint/eslint-plugin-template@13 \
@angular-eslint/schematics@13 \
@angular-eslint/template-parser@13 \
@angular/core@13 @angular/cli@13 @angular/cdk@13 \
@angular/google-maps@13 @angular/compiler-cli@13 \
@angular/language-service@13 @angular/localize@13 \
@angular/animations@13 @angular/common@13 \
@angular/forms@13 @angular/platform-browser@13 \
@angular/router@13 @ng-bootstrap/ng-bootstrap@11 \
shallow-render@13 @angular-devkit/build-angular@13 \
@fortawesome/angular-fontawesome@0.10 angular-shepherd@13 \
ngx-cookie-service@13 @angular/fire@7.2
```

However, individual updates to packages `ts-node` and
`eslint-plugin-jsdoc` were made after the fact.

PER-8854: Upgrade to Angular 13