Fix Shepherd Type Error
In the upgrade to Angular 13, the angular-shepherd library we're using
suddenly became more strongly typed... but it seems like they also
haven't exposed the types the library is expecting? Since we're using
our own interface that matches the `StepOptions` type, let's just cast
it to `any` type when passing it to the `addSteps` method so it behaves
like it did before.

PER-8854: Upgrade to Angular 13