Fix division in Public Gallery Header SCSS
When automatically migrating the usage of the `/` operator in our SCSS,
it did not hit the still-in-progress Gallery code and so we get warnings
when building the project now that it's been merged. Let's just fix it
here so if there are any other Angular 13 related warnings we can see
them more clearly.

PER-8854: Upgrade to Angular 13