Add create account modal to share preview screen
Currently the share preview screen for logged out users features an
uncloseable banner prompting them to log in or create an account. We
want to replace that banner with a banner and modal, both closeable.
This commit adds the modal, and also deactivates the existing banner for
logged out users (there is a separate ticket to remove the code of that
banner).