Fix "Remember Me" Functionality
The "Remember Me" functionality broke at some point in the web-app. This
restores the functionality, by bypassing the check to see if the
"rememberMe" cookie is already set (otherwise it can never be set!).
Additionally, the app is now set to properly stop remembering a username
if the "remember me" checkbox is unchecked.

PER-8958: Remember me checkbox doesn't work