Rename /app/myfiles to /app/private
We are slowly renaming the "My Files" folder to "Private Files". This
changes the url for the Private Root from /app/myfiles/** to
/app/private/**. The core route configuration was changed so
/app/private is now the URL for the private root, and any internal
links/redirects within the app were also changed to point to this new
location.

To make things a bit nicer for people who may have bookmarked URLs
within their archive or may just be hitting a link we missed somewhere
else, a new route guard was made that will redirect any subdirectories
from /app/myfiles/** to /app/private/**. This is used instead of a more
standard redirect to keep the subdirectories in the path intact instead
of just redirecting to /app/private.

PER-8999: Change URL to "/private" instead of "/myfiles"